refactor: remove error pages from pages folder inside components folder
