+++ src/components/BadGatewayPage/BadGatewayPage.tsx
@@ -1,4 +1,4 @@
-import { ErrorPage } from '../../components/shared/ErrorPage/ErrorPage';
+import { ErrorPage } from '../shared/ErrorPage/ErrorPage';
 import image from '../../images/500_large chart.png';
 import { BAD_GATEWAY_SUBTITLE, BAD_GATEWAY_TITLE } from '../../utils/constants';
