+++ src/components/NotFoundPage/NotFoundPage.tsx
@@ -1,4 +1,4 @@
-import { ErrorPage } from '../../components/shared/ErrorPage/ErrorPage';
+import { ErrorPage } from '../shared/ErrorPage/ErrorPage';
 import image from '../../images/404_large_chart.png';
 import { NOT_FOUND_SUBTITLE, NOT_FOUND_TITLE } from '../../utils/constants';
